templates-backend/pyproject.toml
[tool.poetry]
name = "templates"
version = "1.0.0"
description = "Backend of the website Template."
authors = ["Firstname 'Nickname' SURNAME <email>"]
license = "CC BY-NC-SA 4.0"
readme = "../README.rst"

[tool.poetry.dependencies]
python = "^3.11"
psycopg = {extras = ["binary"], version = "^3.1.12"}
alembic = "^1.13.1"
SQLAlchemy = "^2.0.25"
fastapi = {extras = ["all"], version = "^0.111.0"}
strawberry-graphql = "^0.229.0"
strawberry-sqlalchemy-mapper = "==0.3.1"
python-jose = {extras = ["pycryptodome"], version = "^3.3.0"}
requests = "^2.31.0"
uvicorn = "^0.29.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"
sphinxcontrib-mermaid = "^0.9.2"
enum-tools = {extras = ["sphinx"], version = "^0.12.0"}

[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
pytest = "^8.0.1"
pytest-asyncio = "^0.23.5"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
